requirements.txt
flask==2.2.2
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy